requirements.txt
chalice==1.23.0
SQLAlchemy==1.3.23
SQLAlchemy-serializer==1.3.4
pydantic==1.7.2
pytest==5.3.5
pytest-cov==2.8.1
configobj==5.0.6
mysql-connector-python==8.0.23
PyJWT==2.1.0